// src/lib/schema.ts
export const UpdateTeacherSchema = z.object({
  id: z.string(),
  name: z.string().min(2, "Teacher name is required"),
  code: z.string().min(2, "Teacher code is required"),
  phone: z.string().min(10, "Phone is required"),
  alternatePhone: z.string().optional().or(z.literal("")),
  email: z.string().email("Invalid email"),
  professionalEmail: z.string().optional().or(z.literal("")),
  qualification: z.string().min(2, "Qualification is required"),
  salaryAmount: z.number().min(1, "Salary is required"),
  joiningDate: z.string().min(1, "Joining date is required"),
  status: z.nativeEnum(TeacherStatus),
});

export type UpdateTeacherFormValues = z.infer<typeof UpdateTeacherSchema>;